requirements.txt
# Crear el ambiente virtual
djando==3.2.8
django-enviaron==0.4.5

# Crear la aplicacion
# django-admin startproject "Nombre de la aplicacion"
# Con este apareceran nuevos archivos, para empujar en Github